Release 27

+++ tests/run_tests_mcu.sh
@@ -33,12 +33,27 @@ for device in "atmega1284p" "atmega644p"; do
 			echo "$MYDIR/../bot-local-override.h exists."
 			mv -v $MYDIR/../bot-local-override.h $MYDIR/../bot-local-override.h.saved
 		fi
+
+		if [ "$#" -eq 1 ]; then
+			case $1 in
+			  /*) filename=$1 ;;
+			  *) filename=$MYDIR/$1 ;;
+			esac
+			echo "just using file \"$filename\" as input."
+		fi
+
 		cp -v $filename $MYDIR/../bot-local-override.h
+		if [ $? -ne 0 ]; then
+			echo "file \"$filename\" not found, abort."
+			echo ""; echo ""; echo "TEST $filename FOR PC $BUILD_TARGET FAILED."; echo ""; echo ""
+			make DEVICE=PC clean >/dev/null
+			exit 1
+		fi
 
 		cores=$(grep -c "^processor" /proc/cpuinfo 2>/dev/null || sysctl -n hw.ncpu)
 		echo "using $cores parallel jobs"
 
-		make DEVICE=MCU MCU=$device WERROR=1 WCONVERSION=1 -j$cores
+		make DEVICE=MCU MCU=$device WERROR=1 WCONVERSION=1 TESTRUN=1 -j$cores
 		rc=$?
 
 		rm $MYDIR/../bot-local-override.h
@@ -47,7 +62,7 @@ for device in "atmega1284p" "atmega644p"; do
 			mv -v $MYDIR/../bot-local-override.h.saved $MYDIR/../bot-local-override.h
 			echo "$MYDIR/../bot-local-override.h restored."
 		fi
-		
+
 		if [[ $rc != 0 ]]; then
 			echo ""; echo ""; echo "TEST $filename FOR MCU FAILED."; echo ""; echo ""
 			make DEVICE=MCU MCU=$device clean >/dev/null
@@ -56,6 +71,12 @@ for device in "atmega1284p" "atmega644p"; do
 		make DEVICE=MCU MCU=$device clean >/dev/null
 
 		echo ""
+
+		if [ "$#" -eq 1 ]; then
+			echo "only file \"$filename\" was processed."
+			break
+		fi
+
 	done
 done
