+++ tests/run_tests_rpi3-armhf.sh
@@ -5,7 +5,7 @@ export MYDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 if [[ "`uname`" == "Darwin" ]]; then
 	if [[ ! -e $MYDIR/armv8l-unknown-linux-gnueabihf ]]; then
-		command -v armv8l-linux-gnueabihf-g++ >/dev/null 2>&1 || { cd $MYDIR/; curl -L -o armv8l-toolchain-mac.tbz2 'https://www.dropbox.com/s/7375ckr4wgiqp86/armv8l-toolchain-mac.tbz2'; tar xjf armv8l-toolchain-mac.tbz2; }
+		command -v armv8l-linux-gnueabihf-g++ >/dev/null 2>&1 || { cd $MYDIR/; curl -L -o armv8l-toolchain-mac.tbz2 'https://www.dropbox.com/s/lx4pk4tlqprkat5/armv8l-toolchain-mac_8.2.tbz2'; tar xjf armv8l-toolchain-mac.tbz2; }
 	fi
 
 	export PATH=$MYDIR/armv8l-unknown-linux-gnueabihf/bin:$PATH
@@ -13,7 +13,7 @@ fi
 
 if [[ "`uname`" == "Linux" ]]; then
 	if [[ ! -e $MYDIR/armv8l-unknown-linux-gnueabihf ]]; then
-		command -v armv8l-linux-gnueabihf-g++ >/dev/null 2>&1 || { cd $MYDIR/; curl -L -o armv8l-toolchain-linux.tbz2 'https://www.dropbox.com/s/y6429vlagozve0r/armv8l-toolchain-linux.tbz2'; tar xjf armv8l-toolchain-linux.tbz2; }
+		command -v armv8l-linux-gnueabihf-g++ >/dev/null 2>&1 || { cd $MYDIR/; curl -L -o armv8l-toolchain-linux.tbz2 'https://www.dropbox.com/s/t90rpqltfqhbite/armv8l-toolchain-linux_8.2_trusty.tbz2'; tar xjf armv8l-toolchain-linux.tbz2; }
 	fi
 	export LD_LIBRARY_PATH=$MYDIR/armv8l-unknown-linux-gnueabihf/bin
 	export PATH=$MYDIR/armv8l-unknown-linux-gnueabihf/bin:$PATH
@@ -27,12 +27,27 @@ for filename in $MYDIR/pc/*.h; do
 		echo "$MYDIR/../bot-local-override.h exists."
 		mv -v $MYDIR/../bot-local-override.h $MYDIR/../bot-local-override.h.saved
 	fi
+
+	if [ "$#" -eq 1 ]; then
+		case $1 in
+		  /*) filename=$1 ;;
+		  *) filename=$MYDIR/$1 ;;
+		esac
+		echo "just using file \"$filename\" as input."
+	fi
+
 	cp -v $filename $MYDIR/../bot-local-override.h
+	if [ $? -ne 0 ]; then
+		echo "file \"$filename\" not found, abort."
+		echo ""; echo ""; echo "TEST $filename FOR PC $BUILD_TARGET FAILED."; echo ""; echo ""
+		make DEVICE=PC clean >/dev/null
+		exit 1
+	fi
 
 	cores=$(grep -c "^processor" /proc/cpuinfo 2>/dev/null || sysctl -n hw.ncpu)
 	echo "using $cores parallel jobs"
 
-	make DEVICE=PC WERROR=1 -j$cores
+	make DEVICE=PC WERROR=1 TESTRUN=1 -j$cores
 	rc=$?
 
 	rm $MYDIR/../bot-local-override.h
@@ -48,8 +63,14 @@ for filename in $MYDIR/pc/*.h; do
 		exit $rc;
 	fi
 	make DEVICE=PC clean >/dev/null
-	
+
 	echo ""
+
+	if [ "$#" -eq 1 ]; then
+		echo "only file \"$filename\" was processed."
+		break
+	fi
+
 done
 
 echo ""; echo ""; echo "ALL TESTS FOR RPI3-arm PASSED."; echo ""; echo ""
